+++ src/components/format-relative-time.ts
@@ -1,0 +1,1 @@
+export { CpsFormatRelativeTime } from './format-relative-time/format-relative-time.component.js';